point_cloud/requirements.txt
et_xmlfile==2.0.0
gitdb==4.0.11
GitPython==3.1.43
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pyntcloud==0.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
tzdata==2024.2
